+++ creme/__init__.py
@@ -23,6 +23,7 @@ __all__ = [
     'metrics',
     'model_selection',
     'multiclass',
+    'multioutput',
     'naive_bayes',
     'optim',
     'plot',