+++ creme/ensemble/hedge.py
@@ -34,7 +34,7 @@ class HedgeClassifier(base.BinaryClassifier):
             >>> from sklearn import datasets
 
             >>> X_y = stream.iter_sklearn_dataset(
-            ...     load_dataset=datasets.load_breast_cancer,
+            ...     dataset=datasets.load_breast_cancer(),
             ...     shuffle=False
             ... )
             >>> model = compose.Pipeline([