+++ creme/multioutput/__init__.py
@@ -1,0 +1,9 @@
+"""
+Meta-estimators for multi-output classification and regression.
+"""
+from .chain import ClassifierChain
+
+
+__all__ = [
+    'ClassifierChain'
+]